requirements.txt
xmltodict
opencv-python
scipy
tensorboard
requests
pillow
sklearn
wandb
matplotlib
timm
numpy
torch
prettytable
torchsummary
torchinfo
seaborn
pandas
scikit-image
xmltodict
tqdm